✏️ Rename: fix naming conventions of all Svelte component file names to camelCase [#46]
