+++ src/lib/components/LinkIcon.svelte
