+++ src/lib/components/Socials.svelte
