+++ src/lib/components/newsSlider.svelte
@@ -1,6 +1,6 @@
 <!-- © 2022 REVATI -->
 <script lang="ts">
-	import NewsDate from '$lib/components/news_date.svelte';
+	import NewsDate from '$lib/components/NewsDate.svelte';
 
 	import { browser } from '$app/environment';
 	import { NEWS_LIST } from '$lib/data/news';
